FIX: Fixes np.VisibleDeprecationWarning
Was being raised in last step of __call__() function for both Base and
Sampled classes when results for different deltas were stacked into a
single array (since `betas` were always a len-1 array while the other
values were floats). Specifying `dtype=float` removes the warning.

+++ brainsmash/mapgen/base.py
@@ -138,7 +138,7 @@ class Base:
                 res[delta] = self.regress(smvar_perm, self._smvar)
 
             alphas, betas, residuals = np.array(
-                [res[d] for d in self._deltas]).T
+                [res[d] for d in self._deltas], dtype=float).T
 
             # Select best-fit model and regression parameters
             iopt = np.argmin(residuals)
